requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.6
djangorestframework==3.7.1
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
Pygments==2.2.0
pytz==2017.2
requests==2.18.4
uritemplate==3.0.0
urllib3==1.22
uWSGI==2.0.15
django-cors-headers==2.1.0
oauth2-provider==0.0
django-rainbowtests==0.6.0
